compleate backend

+++ backend/controllers/product.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import product from "../models/product.model.js";
 
 export const getProducts = async (req, res) => {
@@ -38,6 +39,13 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({
+      success: false,
+      message: "Invalid product ID",
+    });
+  }
 
   try {
     const deletedProduct = await product.findByIdAndDelete(id);
